Several minor bug fixes/improvements including:
Changed a printed message in the Node.py file for invalid properties into a parse error.

Fixed a bug where some invalid properties would not being detected during parse in PropertyData.py

Fixed a bugs in The SubversionDumpParser that didn't parse property data in a node correctly under certain conditions.

Improved the parsing of the parseHeaderLine and skipEmptyLines method in SVNDumpFilterParser.